Replace `nonnative` with `old`
`Nonnative` used to denote a chipset that was released along with another
generation. This introduced unnecessary clutter and added only redundant
clarity which was provided by the "Socket" row anyway.

As of this commit, `nonnative` has been replaced with `old`, which is only
used for generations which have no chipset of their own, instead
inheriting that of the previous generation.